WIP | WPSEO_Sitemaps_Renderer::encode_url_rfc3986(): add tests
This method, so far, did not have any dedicated tests.

As it is a `protected` method, it cannot be tested directly, but it can be tested via the `get_sitemap()` method.

---

Contains debug code as I think the htmlspecialchars() is done at the wrong time in the code flow.

To investigate:
* Should quotes and such be encoded or escaped ?
* Could esc_url() or esc_xml() be used at the point of creating the output ?

Also:
* Should scheme be added when missing ?